Adding three files. email_code.py sends emails given a proper summary and email list. Template.html is an html template that we edit to send emails with. email_list is a temporary email list used for testing.
