+++ backend/email_list.txt
@@ -1,0 +1,3 @@
+derrickteshiba@g.ucla.edu
+zasaimster@gmail.com
+aszeto35@gmail.com
