--- email_list.txt
@@ -1,3 +1,0 @@
-derrickteshiba@g.ucla.edu
-zasaimster@gmail.com
-aszeto35@gmail.com
