--- template.html
@@ -1,53 +1,0 @@
-<html>
-<head>
-<style>
-    body {
-        background-color: #f1f1f1;
-        width: 100%;
-        height: 100%;
-        font-family: Arial, Helvetica, Sans-serif;
-        margin: 0px;
-        padding: 0px;
-    }
-    .bg {
-        width: 100%;
-        padding: 100%;
-        margin: 0px;
-        padding: 0px;
-        background-color: #f1f1f1;
-        padding: 30px 0px 30px 0px;
-    }
-    .logo {
-        height: 25px;
-    }
-    .content {
-        width: 50%;
-        height: auto;
-        min-height: 200px;
-        border: 1px solid #CCC;
-        background-color: #ffffff;
-        padding: 10px 10px;
-        margin-top: 30px;
-        text-align: justify;
-    }
-</style>
-</head>
-<body>
-    <div class="bg">
-        <table width="100%" border="0" cellspacing="0" cellpadding="0">
-            <tr>
-                <td align="center">
-                    <img src="" class="logo" alt="Logo">
-                </td>
-            </tr>
-            <tr>
-                <td align="center">
-                    <div class="content">
-                    {{ content }}
-                    </div>
-                </td>
-            </tr>
-        </table>
-    </div>
-</body>
-</html>
